[309]: Move Mapper interfaces and associated tests from Util to Isaac folder
