+++ src/main/java/uk/ac/cam/cl/dtg/isaac/mappers/Default.java
@@ -1,4 +1,4 @@
-package uk.ac.cam.cl.dtg.util.mappers;
+package uk.ac.cam.cl.dtg.isaac.mappers;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
