+++ src/main/java/uk/ac/cam/cl/dtg/isaac/mappers/EventMapper.java
@@ -1,4 +1,4 @@
-package uk.ac.cam.cl.dtg.util.mappers;
+package uk.ac.cam.cl.dtg.isaac.mappers;
 
 import java.util.List;
 import org.mapstruct.Mapper;
