+++ src/main/java/uk/ac/cam/cl/dtg/isaac/mappers/UserMapper.java
@@ -1,4 +1,4 @@
-package uk.ac.cam.cl.dtg.util.mappers;
+package uk.ac.cam.cl.dtg.isaac.mappers;
 
 import org.mapstruct.BeanMapping;
 import org.mapstruct.CollectionMappingStrategy;
